+++ system/core/Lang.php
@@ -163,10 +163,11 @@ class CI_Lang {
 	 * Fetches a single line of text from the language array
 	 *
 	 * @param	string	$line		Language line key
+	 * @param array $args Variables to be parsed within string
 	 * @param	bool	$log_errors	Whether to log an error message if the line is not found
 	 * @return	string	Translation
 	 */
-	public function line($line, $log_errors = TRUE)
+	public function line($line, array $args = NULL, $log_errors = TRUE)
 	{
 		$value = isset($this->language[$line]) ? $this->language[$line] : FALSE;
 
@@ -176,9 +177,10 @@ class CI_Lang {
 			log_message('error', 'Could not find the language line "'.$line.'"');
 		}
 
-		return $value;
+		return empty($args)
+						? $value
+						: vsprintf($value, $args);
 	}
-
 }
 
 /* End of file Lang.php */